+++ ant-viewver/src/main/java/com/antourage/weaverlib/screens/base/Repository.kt
@@ -49,7 +49,7 @@ internal class Repository {
                     "Creator $index",
                     "https://djyokoo.com/wp-content/uploads/2018/06/blog-placeholder-800x400.jpg",
                     null, null, null, null, null,
-                    null, null, index.toLong(), null,true, index, true, null
+                    null, null, index.toLong(), null,true, index.toLong(), true, null
                 )
             }
             return object : MockedNetworkBoundResource<List<StreamResponse>>(mockedLiveVideos) {
